backend/requirements.txt
flask
psycopg2-binary
pandas
numpy
scikit-learn
tensorflow
matplotlib
seaborn
jupyter
openpyxl